% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find.R
\name{findCellFrequency}
\alias{findCellFrequency}
\title{findCellFrequency}
\usage{
findCellFrequency(object, min.events = 5)
}
\value{

}
\description{
findCellFrequency
}
\examples{

}
\author{
Dieter Henrik Heiland
}
