% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find.R
\name{findConnections}
\alias{findConnections}
\title{findConnections}
\usage{
findConnections(
  object,
  time_shift = 10,
  distance_max = 0.1,
  pal = viridis::viridis(50),
  parameter_to_col = c("dist", "Time")[1],
  cores = 8,
  Events_select = NULL
)
}
\value{

}
\description{
findConnections
}
\examples{

}
\author{
Dieter Henrik Heiland
}
