% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find.R
\name{findHotspots}
\alias{findHotspots}
\title{findConnections}
\usage{
findHotspots(object, select.freq = 20, cores = 8)
}
\value{

}
\description{
findConnections
}
\examples{

}
\author{
Dieter Henrik Heiland
}
