% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find.R
\name{FindPeacemaker}
\alias{FindPeacemaker}
\title{FindPeacemaker}
\usage{
FindPeacemaker(object, min.sd = 15, verbose = F)
}
\value{

}
\description{
FindPeacemaker
}
\examples{

}
\author{
Dieter Henrik Heiland
}
