% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.R
\name{getVectorfields}
\alias{getVectorfields}
\title{getVectorfields}
\usage{
getVectorfields(
  df,
  NN.file,
  parameter,
  Ram.free = 50000,
  workers = 8,
  dist.spot = 0.01
)
}
\value{

}
\description{
getVectorfields
}
\examples{

}
\author{
Dieter Henrik Heiland
}
