% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_functions.R
\name{initialize,NeuoPhyhysiology-method}
\alias{initialize,NeuoPhyhysiology-method}
\title{Initicate NeuoPhyhysiology object}
\usage{
\S4method{initialize}{NeuoPhyhysiology}(.Object, Data.character)
}
\arguments{
\item{Data.character}{A valid Data.character list}
}
\value{
NeuoPhyhysiology object
}
\description{
Initicate NeuoPhyhysiology object
}
