% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConnections.R
\name{plot_connections}
\alias{plot_connections}
\title{test}
\usage{
plot_connections(
  image,
  poly_connected,
  plot_image_file = NA,
  pal_image = brewer.pal(9, "Reds"),
  alpha_image = 0.8,
  alpha = 1,
  lwd = NA,
  pal = viridis(50)
)
}
\value{

}
\description{
test
}
\examples{

}
\author{
Dieter Henrik Heiland
}
