% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCells}
\alias{plotCells}
\title{plotHotspots}
\usage{
plotCells(
  object,
  cell.keep,
  show.connections = T,
  pt.size = 3,
  color = "red",
  real = F,
  lt.size = 0.1
)
}
\value{

}
\description{
plotHotspots
}
\examples{

}
\author{
Dieter Henrik Heiland
}
