% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotConnections}
\alias{plotConnections}
\title{plotConnections}
\usage{
plotConnections(
  object,
  plot_image_file = NA,
  pal_image = brewer.pal(9, "Reds"),
  alpha_image = 0.8,
  alpha = 1,
  parameter = "Nr_of_con",
  lwd = log(object@Connections$poly_connected$Time + 1)/100,
  pal = viridis::viridis(50)
)
}
\value{

}
\description{
plotConnections
}
\examples{

}
\author{
Dieter Henrik Heiland
}
