% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotHotspots}
\alias{plotHotspots}
\title{plotHotspots}
\usage{
plotHotspots(
  object,
  plot_image_file = NA,
  pal_image = brewer.pal(9, "Reds"),
  alpha_image = 0.8,
  alpha = 0.2,
  factor = 10^7,
  lwd = log(object@Connections$Hotspots$frq + 0.1) * 2,
  pal = colorRampPalette(c("lightgray", "darkgreen", "darkred"))(3)
)
}
\value{

}
\description{
plotHotspots
}
\examples{

}
\author{
Dieter Henrik Heiland
}
