% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotVectorfield}
\alias{plotVectorfield}
\title{plotVectorfield}
\usage{
plotVectorfield(object, real = F, dist.spot = 0.01)
}
\value{

}
\description{
plotVectorfield
}
\examples{

}
\author{
Dieter Henrik Heiland
}
