% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotVectorStream}
\alias{plotVectorStream}
\title{plotVectorfield}
\usage{
plotVectorStream(
  object,
  pt.size = 1,
  pt.alpha = 0.5,
  size.arrow = 1,
  alpha.arrow = 0.5,
  real = F,
  dist.spot = 0.01
)
}
\value{

}
\description{
plotVectorfield
}
\examples{

}
\author{
Dieter Henrik Heiland
}
