% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_functions.R
\name{traces,NeuoPhyhysiology-method}
\alias{traces,NeuoPhyhysiology-method}
\title{Methods}
\usage{
\S4method{traces}{NeuoPhyhysiology}(object)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample from which to extract the content.}
}
\description{
Methods
}
