% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_functions.R
\name{traces}
\alias{traces}
\title{Generics to extract a slots content}
\usage{
traces(object)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample from which to extract the content.}
}
\value{
The respective slots content.
}
\description{
Generics to extract a slots content
}
